tarfs: avoid redundant merge operation for images with only one layer
Avoid redundant merge operation for images with only one layer.

+++ pkg/tarfs/tarfs.go
@@ -492,6 +492,12 @@ func (t *Manager) MergeLayers(ctx context.Context, s storage.Snapshot, storageLo
 		bootstraps = append(bootstraps, metaFilePath)
 	}
 
+	// Merging image with only one layer is a noop, just copy the layer bootstrap as image bootstrap
+	if len(s.ParentIDs) == 1 {
+		metaFilePath := t.layerMetaFilePath(storageLocater(s.ParentIDs[0]))
+		return errors.Wrapf(os.Link(metaFilePath, mergedBootstrap), "create hard link from image bootstrap to layer bootstrap")
+	}
+
 	mergedBootstrapTmp := mergedBootstrap + ".tarfs.tmp"
 	defer os.Remove(mergedBootstrapTmp)
